requirements.txt
streamlit
faiss-cpu
vertexai
google-cloud-aiplatform
google-auth
requests
numpy
pillow

